chore(mainapp): extract upload and download graphs modals
